pyproject.toml
[project]
name = "mcp-wikipedia"
version = "0.1.0"
description = "MCP server for Wikipedia search with time-travel capability"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastmcp>=2.8",
    "httpx",
    "wikitextparser>=0.56.4",
    "pandas",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["wikipedia*"]
